requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.4.0
python-dotenv==1.0.1
pydantic_settings==2.8.1
openai==1.14.3
beautifulsoup4==4.12.3
requests==2.31.0
pymongo==4.8.0